feat: support multi connection for eth wallet
