+++ src/ui/pages/DownloadBackup/__tests__/DownloadBackup.test.tsx
@@ -8,7 +8,7 @@ import { useNavigate } from "react-router-dom";
 import { defaultWalletHookData } from "@src/config/mock/wallet";
 import { downloadBackup } from "@src/ui/ducks/backup";
 import { useAppDispatch } from "@src/ui/ducks/hooks";
-import { useWallet } from "@src/ui/hooks/wallet";
+import { useEthWallet } from "@src/ui/hooks/wallet";
 
 import DownloadBackup from "..";
 
@@ -17,7 +17,7 @@ jest.mock("react-router-dom", () => ({
 }));
 
 jest.mock("@src/ui/hooks/wallet", (): unknown => ({
-  useWallet: jest.fn(),
+  useEthWallet: jest.fn(),
 }));
 
 jest.mock("@src/ui/ducks/backup", (): unknown => ({
@@ -33,7 +33,7 @@ describe("ui/pages/DownloadBackup", () => {
   const mockNavigate = jest.fn();
 
   beforeEach(() => {
-    (useWallet as jest.Mock).mockReturnValue(defaultWalletHookData);
+    (useEthWallet as jest.Mock).mockReturnValue(defaultWalletHookData);
 
     (useAppDispatch as jest.Mock).mockReturnValue(mockDispatch);
