+++ src/ui/pages/Home/useHome.ts
@@ -6,7 +6,7 @@ import { IdentityData } from "@src/types";
 import { useAppDispatch } from "@src/ui/ducks/hooks";
 import { fetchHistory, fetchIdentities, useIdentities } from "@src/ui/ducks/identities";
 import { checkHostApproval } from "@src/ui/ducks/permissions";
-import { useWallet } from "@src/ui/hooks/wallet";
+import { useEthWallet } from "@src/ui/hooks/wallet";
 import { getLastActiveTabUrl } from "@src/util/browser";
 
 export interface IUseHomeData {
@@ -21,7 +21,7 @@ export const useHome = (): IUseHomeData => {
   const dispatch = useAppDispatch();
   const identities = useIdentities();
 
-  const { address, chain, balance } = useWallet();
+  const { address, chain, balance } = useEthWallet();
 
   const refreshConnectionStatus = useCallback(async () => {
     const tabUrl = await getLastActiveTabUrl();
