+++ src/ui/pages/Popup/__tests__/Popup.test.tsx
@@ -11,7 +11,7 @@ import { defaultWalletHookData } from "@src/config/mock/wallet";
 import { Paths } from "@src/constants";
 import { useAppDispatch, useAppSelector } from "@src/ui/ducks/hooks";
 import { usePendingRequests } from "@src/ui/ducks/requests";
-import { useWallet } from "@src/ui/hooks/wallet";
+import { useEthWallet } from "@src/ui/hooks/wallet";
 
 import Popup from "..";
 import { IUsePopupData, usePopup } from "../usePopup";
@@ -26,7 +26,7 @@ jest.mock("@src/ui/ducks/requests", (): unknown => ({
 }));
 
 jest.mock("@src/ui/hooks/wallet", (): unknown => ({
-  useWallet: jest.fn(),
+  useEthWallet: jest.fn(),
 }));
 
 jest.mock("../usePopup", (): unknown => ({
@@ -41,7 +41,7 @@ describe("ui/pages/Popup", () => {
   beforeEach(() => {
     (usePopup as jest.Mock).mockReturnValue(defaultHookData);
 
-    (useWallet as jest.Mock).mockReturnValue(defaultWalletHookData);
+    (useEthWallet as jest.Mock).mockReturnValue(defaultWalletHookData);
 
     (usePendingRequests as jest.Mock).mockReturnValue([{ type: "unknown" }]);
