+++ src/ui/pages/Popup/usePopup.ts
@@ -6,7 +6,7 @@ import { Paths } from "@src/constants";
 import { fetchStatus, useAppStatus } from "@src/ui/ducks/app";
 import { useAppDispatch } from "@src/ui/ducks/hooks";
 import { fetchPendingRequests, usePendingRequests } from "@src/ui/ducks/requests";
-import { useWallet } from "@src/ui/hooks/wallet";
+import { useEthWallet } from "@src/ui/hooks/wallet";
 
 export interface IUsePopupData {
   isLoading: boolean;
@@ -21,7 +21,7 @@ export const usePopup = (): IUsePopupData => {
   const navigate = useNavigate();
 
   const dispatch = useAppDispatch();
-  const { onConnectEagerly } = useWallet();
+  const { onConnectEagerly } = useEthWallet();
   const pendingRequests = usePendingRequests();
   const { isInitialized, isUnlocked, isMnemonicGenerated } = useAppStatus();
   const isShowRequestModal = pendingRequests.length > 0;
